specify bundler version in file

+++ tasks/build.py
@@ -31,6 +31,7 @@ RUBY_VERSION = '.ruby-version'
 GEMFILE = 'Gemfile'
 JEKYLL_CONFIG_YML = '_config.yml'
 HUGO_VERSION = '.hugo-version'
+BUNDLER_VERSION = '.bundler-version'
 
 
 def has_federalist_script():
@@ -182,7 +183,24 @@ def build_jekyll(ctx, branch, owner, repository, site_prefix,
         GEMFILE_PATH = CLONE_DIR_PATH / GEMFILE
         if GEMFILE_PATH.is_file():
             LOGGER.info('Setting up bundler')
-            ctx.run('gem install bundler  --version "<2"')
+            BUNDLER_VERSION_PATH = CLONE_DIR_PATH / BUNDLER_VERSION
+            if BUNDLER_VERSION_PATH.is_file():
+                with BUNDLER_VERSION_PATH.open() as bundler_vers_file:
+                    try:
+                        bundler_version = bundler_vers_file.readline().strip()
+                        # escape-quote the value in case there's anything weird
+                        # in the .bundler-version file
+                        bundler_version = shlex.quote(bundler_version)
+                        regex = r'^[\d]+(\.[\d]+)*$'
+                        bundler_version = re.search(regex, bundler_version).group(0)
+                        if bundler_version:
+                            LOGGER.info('Using bundler version in .bundler-version')
+                            ctx.run(f'gem install bundler  --version "{bundler_version}"')
+                    except Exception:
+                        raise RuntimeError(f'Invalid .bundler-version')
+            else:
+                ctx.run('gem install bundler  --version "<2"')
+
             LOGGER.info('Installing dependencies in Gemfile')
             ctx.run('bundle install')
             jekyll_cmd = 'bundle exec ' + jekyll_cmd
